some Improvements

+++ ACC.h
@@ -33,7 +33,7 @@ public:
 	int getLine(long long* time_nano_seconds, double* x, double* y, double* z);
 
 	//bool setByTime(long long time_nano_seconds);
-	long long entries = 0;
+	size_t entries = 0;
 	~ACC();
 private:
 	std::string fileName_m = "ACC.csv";
